Move ActionGetter to its own file
